+++ src/qusi/light_curve_transforms.py
@@ -14,8 +14,12 @@ def from_light_curve_observation_to_fluxes_array_and_label_array(
     return fluxes, np.array(label, dtype=np.float32)
 
 
-def pair_array_to_tensor(arrays: tuple[npt.NDArray[np.float32], npt.NDArray[np.float32]]) -> (Tensor, Tensor):
-    return torch.tensor(arrays[0]), torch.tensor(arrays[1])
+def pair_array_to_tensor(
+    arrays: tuple[npt.NDArray[np.float32], npt.NDArray[np.float32]],
+) -> (Tensor, Tensor):
+    return torch.tensor(arrays[0], dtype=torch.float32), torch.tensor(
+        arrays[1], dtype=torch.float32
+    )
 
 
 def randomly_roll_elements(example: np.ndarray) -> np.ndarray:
